requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
h11==0.14.0
idna==3.4
logger==1.4
loguru==0.6.0
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.3.3
schematics==2.1.1
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
win32-setctime==1.1.0
